Add directive for new home tab
